+++ src/ascnhalf-world/Item.h
@@ -135,7 +135,7 @@ public:
 	
 	//! DB Serialization
 	void LoadFromDB( Field *fields, Player* plr, bool light );
-	void SaveToDB( int8 containerslot, int8 slot, bool firstsave, QueryBuffer* buf );
+	void SaveToDB( int16 containerslot, int16 slot, bool firstsave, QueryBuffer* buf );
 	bool LoadAuctionItemFromDB( uint64 guid );
 	void DeleteFromDB();
 	bool IsEligibleForRefund();
